+++ automig/__main__.py
@@ -45,10 +45,10 @@ alter table automigrate_meta add column if not exists opaque bool;
 alter table automigrate_meta add column if not exists automig_version text;
 """
 
-def main():
-  args = build_parser().parse_args()
+def main_inner(args):
   rev_tuple = githelp.parse_rev_to_tuple(args.ref)
-  print(f'-- changeset created from {args} at {datetime.now()}')
+  lines = []
+  lines.append(f'-- changeset created from {args} at {datetime.now()}')
   shas = []
   manual_mig = {}
   if os.path.exists('.manualmig.yml'):
@@ -58,13 +58,13 @@ def main():
     if args.opaque:
       raise ValueError("don't pass --opaque with --initial")
     shas.append(None)
-    print(PREAMBLE)
+    lines.append(PREAMBLE)
     commit = git.Repo(search_parent_directories=True).commit(rev_tuple[0])
     shas.append(commit.hexsha)
     tree = commit.tree
     assert len(rev_tuple) == 1, "can't pass a sha range for --initial"
     for contents in githelp.get_streams(tree, pathlib.Path.cwd() / pathlib.Path(args.glob)):
-      print(contents.decode())
+      lines.append(contents.decode())
   else:
     assert len(rev_tuple) == 2, "must pass a sha range or set --initial"
     # todo: look up rev_tuple[0] so this isn't a short sha or HEAD~1 or something
@@ -77,19 +77,19 @@ def main():
       if remaining:
         raise ValueError('errors not overridden in .manualmig.yml', remaining)
     if args.update_meta:
-      print(POSTGRES_PREAMBLE)
+      lines.append(POSTGRES_PREAMBLE)
     for sha, tables in changes.items():
       if not tables:
         continue
       shas.append(sha)
       for table, stmts in tables.items():
-        print(f'-- changes for {sha[:7]}.{table}')
-        for stmt in stmts:
-          print(stmt)
+        lines.append(f'-- changes for {sha[:7]}.{table}')
+        lines.extend(stmts)
   for sha1, sha2 in zip(shas[:-1], shas[1:]):
     # todo: hard w/out depending on a sql driver but do some kind of escaping here
     sha1 = f"'{sha1}'" if sha1 is not None else 'NULL'
-    print(f"insert into automigrate_meta (fromsha, sha, automig_version, opaque) values ({sha1}, '{sha2}', '{read_version()}', {['false', 'true'][args.opaque]});")
+    lines.append(f"insert into automigrate_meta (fromsha, sha, automig_version, opaque) values ({sha1}, '{sha2}', '{read_version()}', {['false', 'true'][args.opaque]});")
+  return '\n'.join(lines)
 
 if __name__ == '__main__':
-  main()
+  print((main_inner(build_parser().parse_args())))